Modify CreateMovieRequest to use MinimumAgeRating enum instead of int
Agent-Logs-Url: https://github.com/boonsboos/ticket-play/sessions/c3dc6e68-79ea-44f3-bfbe-7ec650d534e7

+++ ConnectPlay.TicketPlay.UI/App/Pages/Manager/CreateMovie.razor.cs
@@ -38,7 +38,7 @@ public partial class CreateMovie(IMovieApi movieApi) : TranslatableComponent
                 ReleaseDate = DateOnly.FromDateTime(model.ReleaseDate!.Value),
                 PosterUrl = posterUri,
                 Language = model.Language!,
-                MinimumAge = (int)model.MinimumAge!.Value,
+                MinimumAge = model.MinimumAge!.Value,
                 Genre = model.Genre switch
                 {
                     Genre.Thriller => "thriller",
